third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-content-type-charset-expected.txt

CONSOLE ERROR: EventSource's response has a charset ("windows-1251") that is not UTF-8. Aborting the connection.
CONSOLE ERROR: EventSource's response has a MIME type ("text/event-stream-foobar") that is not "text/event-stream". Aborting the connection.
[Worker] Test EventSource with an event-stream with a Content-Type with a charset is still recognized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-content-type-charset.js
PASS [Worker] es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3DUTF-8") did not throw exception.
[Worker] Content type: text/event-stream; charset=UTF-8
PASS [Worker] !!es.sawError is false
PASS [Worker] es.sawMessage is true
PASS [Worker] es.sawOpen is true
PASS [Worker] es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3Dwindows-1251") did not throw exception.
[Worker] Content type: text/event-stream; charset=windows-1251
PASS [Worker] es.sawError is true
PASS [Worker] !!es.sawMessage is false
PASS [Worker] !!es.sawOpen is false
PASS [Worker] es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3Dutf-8") did not throw exception.
[Worker] Content type: text/event-stream; charset=utf-8
PASS [Worker] !!es.sawError is false
PASS [Worker] es.sawMessage is true
PASS [Worker] es.sawOpen is true
PASS [Worker] es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3D%22UTF-8%22") did not throw exception.
[Worker] Content type: text/event-stream; charset="UTF-8"
PASS [Worker] !!es.sawError is false
PASS [Worker] es.sawMessage is true
PASS [Worker] es.sawOpen is true
PASS [Worker] es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream-foobar%3B") did not throw exception.
[Worker] Content type: text/event-stream-foobar;
PASS [Worker] es.sawError is true
PASS [Worker] !!es.sawMessage is false
PASS [Worker] !!es.sawOpen is false
PASS successfullyParsed is true

TEST COMPLETE
